# Speed up Dirac submission