--- a/python/GangaDirac/Lib/Backends/DiracBase.py
+++ b/python/GangaDirac/Lib/Backends/DiracBase.py
@@ -157,7 +157,7 @@ class DiracBase(IBackend):
         dirac_cmd = """execfile(\'%s\')""" % dirac_script
 
         try:
-            result = execute(dirac_cmd, cred_req=self.credential_requirements, return_raw_dict = True)
+            result = execute(dirac_cmd, cred_req=self.credential_requirements)
         except GangaDiracError as err:
 
             err_msg = 'Error submitting job to Dirac: %s' % str(err)