Fixes #1129 .
This just puts each submission into the same process. Currently goes from 130s to 23s (of which it looks like 16s are other overheads so only 7s for the subjob submission) for 23 subjobs.

Needs tidying and potentially new script templates although that will also affect RTHandlers.