@mesmith75 2017-10-02
So I think this is ready. I decided against redoing the RTHandlers and script templates. Currently the job script therefore has a repetition of 
```python
from DIRAC.Interfaces.API.Dirac import Dirac
from DIRAC.Interfaces.API.Job import Job
dirac = Dirac()
```
or similar for the LHCb apps. The subsequent loss of time should be minimal but it would be nice to get rid of it.