@rob-c 2017-10-02
@mesmith75 Would it be insane to have a method which removes all repeated lines from the dirac script before submitting it as a work around to declaring Dirac over and over? (This could be a dangerous idea but I don't know if it might help)
The speed improvement in submitting jobs this way looks impressive.

Looking forward to trying this in the future 👍, only minor comment is are the `stripProxy` calls needed in code within the backend?