@mesmith75 2017-10-02
@rob-c Sure, I've added in a bit to get rid of the repeated imports etc - just needed to discover regex.

This is ready now.