@mesmith75 2017-10-03
Thanks for the review - I have addressed those issues and tidied a bit further. If there are no objections I'll merge this today.